Ids zu int geändert

+++ src/main/java/de/hhu/propra/exambyte/domain/model/nutzerin/NutzerIn.java
@@ -1,24 +1,23 @@
 package de.hhu.propra.exambyte.domain.model.nutzerin;
 
 import org.springframework.data.annotation.Id;
-import org.springframework.data.annotation.PersistenceCreator;
 
 public class NutzerIn {
 
     @Id
-    private final Long id;
+    private final int id;
     private final String username;
 
-    public NutzerIn(Long id, String username) {
+    public NutzerIn(int id, String username) {
         this.id = id;
         this.username = username;
     }
 
-    public NutzerIn withId(Long id) {
+    public NutzerIn withId(int id) {
         return new NutzerIn(id, username);
     }
 
-    public Long id() {
+    public int id() {
         return id;
     }
